Revert "Coloquei @login_required nas funções."
This reverts commit 4764c67f53b66ae759d3b4f2b8493793ea273016.

+++ sistema/views.py
@@ -21,64 +21,91 @@ def teste(request):
 
 
 @login_required
-def projeto_lista(request):
-    lista_projetos = Projeto.objects.all()
-
-    return render(request, 'projeto_lista.html', locals())
-
-@login_required
 def projeto_cadastrar(request):
-    novo_projeto = ProjetoForm()
-    if request.method == 'POST':
-        novo_projeto = ProjetoForm(request.POST)
-        print novo_projeto
-        if novo_projeto.is_valid():
-            projeto = novo_projeto.save(commit=False)
-            projeto.save()
-            return HttpResponseRedirect('/projeto_perfil/%s' % str(projeto.id))
+    membro = request.user.membro
+    projeto = Projeto.objects.all()
+    cargo = Cargo.objects.filter(projeto = projeto, cargo__iexact = 'coordenador')
+    if cargo:
+        novo_projeto = ProjetoForm()
+        if request.method == 'POST':
+            novo_projeto = ProjetoForm(request.POST)
+            print novo_projeto
+            if novo_projeto.is_valid():
+                projeto = novo_projeto.save(commit=False)
+                projeto.save()
+                return HttpResponse('<script>alert("Projeto cadastrado com sucesso"); location.replace("/projeto_perfil/%s")</script>' % str(projeto.id))
+    else:
+        return HttpResponse('<script>alert("Você não tem permissão para cadastrar um projeto."); history.back()</script>')
 
     lista_projetos = Projeto.objects.all()
 
     return render(request, 'projeto_cadastrar.html', locals())
+
+@login_required
+def projeto_lista(request):
+    lista_projetos = Projeto.objects.all()
+    lista_geral = []
+
+    for projeto in lista_projetos:
+        gerente_lista = Cargo.objects.filter(projeto = projeto, cargo__iexact = 'gerente')
+        if len(gerente_lista) > 0:
+            gerente = Cargo.objects.get(projeto = projeto, cargo__iexact = 'gerente')
+            lista_geral.append([projeto, gerente.membro.nome + ' ' + gerente.membro.sobrenome, gerente.nucleo.nome])
+        else:
+            lista_geral.append([projeto, 'Gerente não detectado.', 'Núcleo não detectado.'])
+    return render(request, 'projeto_lista.html', locals())
 
 @login_required
 def projeto_editar(request, projeto_id):
     projeto = Projeto.objects.get(id=projeto_id)
     novo_projeto = ProjetoForm(instance=projeto)
-
-    if request.method == 'POST':
-        novo_projeto = ProjetoForm(request.POST, instance=projeto)
-        if novo_projeto.is_valid():
-            projeto = novo_projeto.save()
-            return HttpResponseRedirect('/projeto_perfil/%s' % str(projeto.id))
+    membro = request.user.membro
+    projeto2 = Projeto.objects.all()
+    cargo = Cargo.objects.filter(projeto = projeto2, cargo__iexact = 'coordenador')
+    if cargo:
+        if request.method == 'POST':
+            novo_projeto = ProjetoForm(request.POST, instance=projeto)
+            if novo_projeto.is_valid():
+                projeto = novo_projeto.save()
+                return HttpResponse('''
+                    <script>alert("Projeto editado com sucesso");
+                    location.replace("/projeto_perfil/%s")
+                    </script>''' % str(projeto.id))
+        else:
+            return HttpResponse('<script>alert("Você não tem permissão para editar um projeto."); history.back()</script>')
 
     lista_projetos = Projeto.objects.all()
 
     return render(request, 'projeto_cadastrar.html', locals())
 
+
 @login_required
 def projeto_perfil(request, projeto_id):
     projeto = Projeto.objects.get(id=projeto_id)
     nome_projeto = projeto.nome
-
     lista_gerente = Cargo.objects.filter(projeto = projeto, cargo__iexact = 'gerente')
     if not lista_gerente:
         gerente_erro = 'Não há nenhum gerente alocado para esse projeto'
-
     analistas = Cargo.objects.filter(projeto = projeto)
     if not analistas:
         analistas_erro = 'Não há nenhum membro alocado para esse projeto'
 
     return render(request, 'projeto_perfil.html', locals())
 
+
 @login_required
 def projeto_deletar(request, projeto_id):
-    projeto = Projeto.objects.get(id=projeto_id)
-    projeto.delete()
+    membro = request.user.membro
+    projeto = Projeto.objects.all()
+    cargo = Cargo.objects.filter(projeto = projeto, cargo__iexact = 'coordenador')
+    if cargo:
+        projeto = Projeto.objects.get(id=projeto_id)
+        projeto.delete()
+        return HttpResponse('<script>alert("Projeto deletado com sucesso"); location.replace("/projeto_lista/")</script>')
+    else:
+        return HttpResponse('<script>alert("Você não tem permissão para cadastrar um projeto."); history.back()</script>')
 
-    return HttpResponseRedirect("/projeto_lista/")
 
-@login_required
 def cadastra_usuario(request):
     cadastro = True
     membro_form = MembroForm()
@@ -158,7 +185,7 @@ def perfil_usuario(request, usuario_id):
         lista_cargos = False
     return render(request, 'perfil_usuario.html', locals())
 
-@login_required
+
 def lista_usuario(request):
     id_atual = request.user.membro.id
     coordenador = False
@@ -180,7 +207,7 @@ def lista_usuario(request):
 
     return render(request, 'lista_usuario.html', locals())
 
-@login_required
+
 def cadastrar_nucleo(request):
     nucleo_form = NucleoForm()
     if request.method == "POST":
@@ -192,7 +219,6 @@ def cadastrar_nucleo(request):
     texto = "Cadastro de um novo núcleo"
     return render(request, 'cadastrar_nucleo.html', locals())
 
-@login_required
 def atualizar_nucleo(request, nucleo_id):
     nucleo = Nucleo.objects.get(id = nucleo_id)
     nucleo_form = NucleoForm(instance = nucleo)
@@ -205,18 +231,16 @@ def atualizar_nucleo(request, nucleo_id):
     texto = "Atualizar núcleo"
     return render(request, 'cadastrar_nucleo.html', locals())
 
-@login_required
 def apagar_nucleo(request, nucleo_id):
     nucleo = Nucleo.objects.get(id = nucleo_id)
     nucleo.delete()
     return HttpResponse('<script>location.replace("/ver_nucleos/")</script>')
 
-@login_required
 def ver_nucleos(request):
     lista_nucleo = Nucleo.objects.all()
     return render(request, 'ver_nucleos.html', locals())
 
-@login_required
+
 def cadastra_cargo(request, usuario_id):
     membro = Membro.objects.get(id=usuario_id)
     cargo_form = CargoForm()
@@ -229,20 +253,26 @@ def cadastra_cargo(request, usuario_id):
             return HttpResponse('<script>alert("Cargo cadastrado com sucesso"); history.back()</script>')
     return render(request, 'cadastra_cargo.html', locals())
 
-@login_required
+
 def cadastra_cargo2(request, projeto_id):
+    membro = request.user.membro
     projeto = Projeto.objects.get(id=projeto_id)
+    projeto2 = Projeto.objects.all()
     cargo_form = CargoForm2()
-    if request.method == 'POST':
-        cargo_form = CargoForm2(request.POST)
-        if cargo_form.is_valid():
-            cargo = cargo_form.save(commit = False)
-            cargo.projeto = projeto
-            cargo.save()
-            return HttpResponse('<script>alert("Cargo cadastrado com sucesso"); location.replace("/projeto_perfil/%s")</script>' %str(projeto.id))
+    cargo = Cargo.objects.filter(projeto = projeto2, cargo__iexact = 'coordenador')
+    if cargo:
+        if request.method == 'POST':
+            cargo_form = CargoForm2(request.POST)
+            if cargo_form.is_valid():
+                cargo = cargo_form.save(commit = False)
+                cargo.projeto = projeto
+                cargo.save()
+                return HttpResponse('<script>alert("Membro adicionado com sucesso"); location.replace("/projeto_perfil/%s")</script>' %str(projeto.id))
+    else:
+        return HttpResponse('<script>alert("Você não tem permissão para adicionar um membro."); history.back()</script>')
     return render(request, 'cadastra_cargo.html', locals())
 
-@login_required
+
 def deleta_cargo(request, cargo_id):
     cargo = Cargo.objects.get(id=cargo_id)
     cargo.delete()
